demo/pyproject.toml
[tool.poetry]
name = "demo"
version = "0.1.0"
description = ""
authors = ["Abhijit Gupta <avgupta456@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
SQLAlchemy = "^1.4.40"
python-dotenv = "^0.21.1"
sqlalchemy-cockroachdb = "^1.4.4"
psycopg2 = "^2.9.5"
openai = "^0.26.5"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
